fix: resolve relative import paths after moving app directory
