+++ src/app/(drawer)/(tabs)/index.tsx
@@ -1,4 +1,4 @@
-import HomeScreen from '../../../src/screens/HomeScreen'
+import HomeScreen from '../../../screens/HomeScreen'
 
 export default function DayScreen() {
     return <HomeScreen viewMode="day" />
