+++ src/app/(drawer)/(tabs)/month.tsx
@@ -1,4 +1,4 @@
-import HomeScreen from '../../../src/screens/HomeScreen'
+import HomeScreen from '../../../screens/HomeScreen'
 
 export default function MonthScreen() {
     return <HomeScreen viewMode="month" />
