+++ src/app/(drawer)/(tabs)/week.tsx
@@ -1,4 +1,4 @@
-import HomeScreen from '../../../src/screens/HomeScreen'
+import HomeScreen from '../../../screens/HomeScreen'
 
 export default function WeekScreen() {
     return <HomeScreen viewMode="week" />
