+++ src/app/geofence-setup.tsx
@@ -1,4 +1,4 @@
-import GeofenceSetupScreen from '../src/screens/menu/GeofenceSetupScreen'
+import GeofenceSetupScreen from '../screens/menu/GeofenceSetupScreen'
 
 export default function GeofenceSetupRoute() {
     return <GeofenceSetupScreen />
