+++ src/app/language-settings.tsx
@@ -1,4 +1,4 @@
-import LanguageSettingsScreen from '../src/screens/settings/LanguageSettingsScreen'
+import LanguageSettingsScreen from '../screens/settings/LanguageSettingsScreen'
 
 export default function LanguageSettingsRoute() {
     return <LanguageSettingsScreen />
