+++ src/app/licenses.tsx
@@ -1,4 +1,4 @@
-import LicensesScreen from '../src/screens/menu/LicensesScreen'
+import LicensesScreen from '../screens/menu/LicensesScreen'
 
 export default function LicensesRoute() {
     return <LicensesScreen />
