+++ src/app/nfc-setup.tsx
@@ -1,4 +1,4 @@
-import NFCSetupScreen from '../src/screens/menu/NFCSetupScreen'
+import NFCSetupScreen from '../screens/menu/NFCSetupScreen'
 
 export default function NFCSetupRoute() {
     return <NFCSetupScreen />
